added the student collection
