+++ src/app.ts
@@ -1,7 +1,7 @@
 import express, { Application, Request, Response } from 'express';
 import cors from 'cors';
-import { UserRoutes } from './app/modules/users/user.route';
 import { ResultsRoutes } from './app/modules/results/results.route';
+import {StudentRoutes} from "./app/modules/student/student.route";
 
 const app: Application = express();
 
@@ -10,7 +10,7 @@ const app: Application = express();
 app.use(express.json());
 app.use(cors());
 
-app.use('/api', UserRoutes);
+app.use('/api', StudentRoutes);
 app.use('/api', ResultsRoutes);
 
 app.get('/', (req: Request, res: Response) => {
